Migrate remaining components to TypeScript.
