+++ src/components/ModelTableList/ModelTableList.tsx
@@ -1,10 +1,17 @@
+import { Filters } from "store/juju/utils/models";
+
 import StatusGroup from "./StatusGroup";
 import CloudGroup from "./CloudGroup";
 import OwnerGroup from "./OwnerGroup";
 
 import "./_model-table-list.scss";
 
-export default function ModelTableList({ filters, groupedBy }) {
+type Props = {
+  groupedBy: string;
+  filters: Filters;
+};
+
+export default function ModelTableList({ filters, groupedBy }: Props) {
   switch (groupedBy) {
     case "status":
     default:
